Move recurring tasks back to default bucket when done
Closes #347

The idea behind this is a kanban layout like this:

    | Backlog | In Progress | Done |

A recurring task is also moved into `In Progress`, however when it's
closed, it should be moved back into the default column (Backlog).

This patch effectively implements that by changing the bucket ID in the
saved view event hook for recurring tasks that were closed in the update
action. The frontend counterpart gets a condition to detect this case
and move the card to the right column in the UI.

This patch is a POC to see if this goes into the right direction, most
notably there's one part of the feature missing: I mostly care about
the Ctrl-Click case, the "move to 'Done'" case doesn't work. I'm happy
to implement that if the rest of the patch is fine for upstream.